backend/requirements.txt
flask==2.3.3
flask-cors==3.0.10
python-dotenv==0.19.1
pymysql==1.0.2
sqlalchemy==1.4.23
flask-sqlalchemy==3.0.5
jinja2==3.1.2
openai>=1.0.0
gunicorn==20.1.0
werkzeug==2.3.7 